Cleaning up various minor issues

+++ SHE_Validation_CTI/python/SHE_Validation_CTI/table_formats/regression_results.py
@@ -23,10 +23,8 @@ __updated__ = "2021-08-26"
 
 from collections import OrderedDict
 
-from SHE_PPT.constants.shear_estimation_methods import ShearEstimationMethods
 from SHE_PPT.logging import getLogger
 from SHE_PPT.table_utility import SheTableFormat, SheTableMeta
-
 
 FITS_VERSION = "8.0"
 FITS_DEF = "she.regressionResults"
@@ -49,16 +47,15 @@ class SheRegressionResultsMeta(SheTableMeta):
     bin_limit_max: str = "BIN_MAX"
 
     def __init__(self):
-
         # Store the less-used comments in a dict
-        super().__init__(comments=OrderedDict(((self.fits_version, None),
-                                               (self.fits_def, None),
-                                               (self.product_type, "Whether this is a test of an observation (OBS) " +
-                                                "or exposure (EXP)"),
-                                               (self.test_case, None),
-                                               (self.bin_limit_min, None),
-                                               (self.bin_limit_max, None),
-                                               )))
+        super().__init__(comments = OrderedDict(((self.fits_version, None),
+                                                 (self.fits_def, None),
+                                                 (self.product_type, "Whether this is a test of an observation (OBS) " +
+                                                  "or exposure (EXP)"),
+                                                 (self.test_case, None),
+                                                 (self.bin_limit_min, None),
+                                                 (self.bin_limit_max, None),
+                                                 )))
 
 
 class SheRegressionResultsFormat(SheTableFormat):
@@ -68,14 +65,13 @@ class SheRegressionResultsFormat(SheTableFormat):
     """
 
     def __init__(self):
-
         # Get the metadata (contained within its own class)
         super().__init__(SheRegressionResultsMeta())
 
         # Table column labels
 
-        self.index = self.set_column_properties(f"INDEX", dtype=int, fits_dtype=">i2")
-        self.method = self.set_column_properties(f"METHOD", dtype=str, fits_dtype="A", length=9)
+        self.index = self.set_column_properties(f"INDEX", dtype = int, fits_dtype = ">i2")
+        self.method = self.set_column_properties(f"METHOD", dtype = str, fits_dtype = "A", length = 9)
 
         self.weight = self.set_column_properties(f"WEIGHT")
         self.slope = self.set_column_properties(f"M")
